+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Dev: Do a pretty major refactor of the Settings classes. List settings (e.g. highlights) are most heavily modified, and should have an extra eye kept on them. (#4775)
 - Dev: Remove `boost::noncopyable` use & `boost_random` dependency. (#4776)
 - Dev: Fix clang-tidy `cppcoreguidelines-pro-type-member-init` warnings. (#4426)
+- Dev: Immediate layout for invisible `ChannelView`s is skipped. (#4811)
 
 ## 2.4.5